+++ src/audio/noise/mod.rs
@@ -1,0 +1,3 @@
+mod description;
+
+pub use description::NoiseDescription;
