+++ src/audio/pulse/sweep/direction.rs
@@ -1,0 +1,5 @@
+#[derive(PartialEq, Copy, Clone)]
+pub enum SweepDirection {
+    Add,
+    Sub,
+}
