+++ src/audio/pulse/wave_pattern_duty.rs
