+++ src/audio/volume_envelope/mod.rs
@@ -1,0 +1,5 @@
+mod description;
+mod direction;
+
+pub use self::description::VolumeEnvelopeDescription;
+pub use self::direction::VolumeEnvelopeDirection;
