+++ src/audio/wave/mod.rs
@@ -1,0 +1,5 @@
+mod description;
+mod output_level;
+
+pub use self::description::WaveDescription;
+pub use self::output_level::WaveOutputLevel;
